[VLM] Extracted inputs embedder
